FINALLY 0 KOs

+++ src/main/java/rinha/backendq1/models/TransactionRequest.java
@@ -1,4 +1,4 @@
 package rinha.backendq1.models;
 
-public record TransactionRequest(Integer valor, String tipo, String descricao) {
+public record TransactionRequest(Double valor, String tipo, String descricao) {
 };
